+++ tw-base-utils/src/test/java/com/transferwise/common/baseutils/UuidUtilsTest.java
@@ -9,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.google.common.collect.Ordering;
 import com.transferwise.common.baseutils.clock.TestClock;
+import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Arrays;
@@ -49,21 +50,26 @@ class UuidUtilsTest extends BaseTest {
   }
 
   @Test
-  void uuidv7IsGrowingOverTime() {
+  void timePrefixedUuidIsVersion7() {
+    UUID uuid = UuidUtils.generateTimePrefixedUuid();
+    assertEquals(7, uuid.version());
+  }
+
+  @Test
+  void timePrefixedUuidIsGrowingOverTime() {
     TestClock clock = TestClock.createAndRegister();
 
     int n = 100;
 
     UUID[] uuids = new UUID[n];
     for (int i = 0; i < n; i++) {
-      uuids[i] = UuidUtils.generateUuidv7();
+      uuids[i] = UuidUtils.generateTimePrefixedUuid();
       clock.tick(Duration.ofMillis(2));
     }
 
     long previousTime = -1;
     for (int i = 0; i < n; i++) {
       System.out.println("Testing " + uuids[i]);
-      assertEquals(7, uuids[i].version());
       long time = uuids[i].getMostSignificantBits() >>> (64 - 48);
 
       if (previousTime != -1) {
@@ -77,8 +83,59 @@ class UuidUtilsTest extends BaseTest {
   }
 
   @Test
+  void deterministicTimePrefixedUuidIsVersion8() {
+    UUID uuid = UuidUtils.generateDeterministicTimePrefixedUuid(new byte[0]);
+    System.out.println(uuid);
+    assertEquals(8, uuid.version());
+  }
+
+  @Test
+  void deterministicTimePrefixedUuidIsGrowingOverTime() {
+    TestClock clock = TestClock.createAndRegister();
+
+    int n = 100;
+
+    UUID[] uuids = new UUID[n];
+    for (int i = 0; i < n; i++) {
+      uuids[i] = UuidUtils.generateDeterministicTimePrefixedUuid(new byte[0]);
+      clock.tick(Duration.ofMillis(2));
+    }
+
+    long previousTime = -1;
+    for (int i = 0; i < n; i++) {
+      System.out.println("Testing " + uuids[i]);
+      long time = uuids[i].getMostSignificantBits() >>> (64 - 48);
+
+      if (previousTime != -1) {
+        assertTrue(previousTime < time);
+      }
+      previousTime = time;
+    }
+
+    assertTrue(Ordering.natural().isOrdered(Arrays.asList(uuids)));
+    assertEquals(n, Set.of(uuids).size());
+  }
+
+  @Test
+  void deterministicTimePrefixedUuidIsFullyDeterministic() {
+    long timestamp = System.currentTimeMillis();
+    byte[] seed = "Ben was here".getBytes(StandardCharsets.UTF_8);
+    UUID uuid1 = UuidUtils.generateDeterministicTimePrefixedUuid(timestamp, seed);
+    UUID uuid2 = UuidUtils.generateDeterministicTimePrefixedUuid(timestamp, seed);
+    assertEquals(uuid1, uuid2);
+  }
+
+  @Test
+  void deterministicTimePrefixedUuidChangesWithDifferentData() {
+    long timestamp = System.currentTimeMillis();
+    UUID uuid1 = UuidUtils.generateDeterministicTimePrefixedUuid(timestamp, "Ben was here".getBytes(StandardCharsets.UTF_8));
+    UUID uuid2 = UuidUtils.generateDeterministicTimePrefixedUuid(timestamp, "Ben was not here".getBytes(StandardCharsets.UTF_8));
+    assertNotEquals(uuid1, uuid2);
+  }
+
+  @Test
   void constantCanBeAddedToUuid() {
-    var uuid = UuidUtils.generatePrefixCombUuid();
+    var uuid = UuidUtils.generateTimePrefixedUuid();
     var uuidWithConstant = UuidUtils.add(uuid, 11111);
 
     assertNotEquals(uuidWithConstant, uuid);
@@ -94,13 +151,13 @@ class UuidUtilsTest extends BaseTest {
 
     UUID[] uuids = new UUID[n];
     for (int i = 0; i < n; i++) {
-      uuids[i] = UuidUtils.add(UuidUtils.generatePrefixCombUuid(), ThreadLocalRandom.current().nextInt());
+      uuids[i] = UuidUtils.add(UuidUtils.generateTimePrefixedUuid(), ThreadLocalRandom.current().nextInt());
       clock.tick(Duration.ofMillis(2));
     }
 
     long previousTime = -1;
     for (int i = 0; i < n; i++) {
-      long time = uuids[i].getMostSignificantBits() >>> (64 - 38);
+      long time = uuids[i].getMostSignificantBits() >>> (64 - 48);
 
       if (previousTime != -1) {
         assertTrue(previousTime < time);
@@ -108,7 +165,7 @@ class UuidUtilsTest extends BaseTest {
       previousTime = time;
 
       System.out.println(uuids[i]);
-      assertEquals(4, uuids[i].version());
+      assertEquals(7, uuids[i].version());
     }
 
     assertTrue(Ordering.natural().isOrdered(Arrays.asList(uuids)));